+++ impl/src/main/java/com/tech/dimefresh/service/dto/NewsDataProjection.java
@@ -14,6 +14,6 @@ public record NewsDataProjection(
         String name,
 
         Instant publishedAt,
-        Integer likedAccounts
+        Long likedAccounts
 ) {
 }
